+++ CHANGELOG.md
@@ -4,7 +4,11 @@
 
 ## XX January 2025 (0.9.4)
 
-### DIDComm Library
+### All (0.9.4)
+
+* Rand crate updated from 0.8.x to 0.9.x
+
+### DIDComm Library (0.9.4)
 
 * Cleaned up unneeded lifetime parameters
 * Changed how DID Document Verification Methods are discovered, more robust algorithm used
@@ -22,12 +26,17 @@
 * FIX: Mediator Administration generates a client side error when no Admin DID is selected when removing Admin Accounts
 * FIX: Double hashing of DID's on admin_add, refactored so now only uses SHA256 hashed DID's
 * FEATURE: Added AccountChangeType to the mediator account-management protocol
+* FIX/FEATURE: Mediator will detect when forwarding a message to itself.
+  * When a forward to itself is detected, it will block the forward and deliver locally
+  * Added configurtion for other local mediator DID's that you want to block forwarding towards
 
 ### SDK (0.9.4)
 
 * FIX: If ProblemReport had no args, deserializing would fail as no args field. Now defaults to empty array correctly
 * TEST: Added ProblemReport tests to check for empty args and serialization/deserialization
 * FEATURE: Added AccountChangeType to the mediator account-management protocol
+
+---
 
 ## 18th January 2025 (0.9.2)